Move stringifier registry to upper level
