+++ client/src/test/java/org/spine3/base/IdentifiersShould.java
@@ -39,7 +39,6 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertTrue;
 import static org.spine3.base.Identifiers.CommandIdToStringConverter;
-import static org.spine3.base.Identifiers.ConverterRegistry;
 import static org.spine3.base.Identifiers.EMPTY_ID;
 import static org.spine3.base.Identifiers.NULL_ID;
 import static org.spine3.base.Identifiers.idToAny;
